+++ thoth-export-server/src/json/json_thoth.rs
@@ -111,6 +111,7 @@ mod tests {
             imprint: WorkImprint {
                 imprint_name: "OA Editions Imprint".to_string(),
                 imprint_url: None,
+                crossmark_doi: None,
                 publisher: WorkImprintPublisher {
                     publisher_name: "OA Editions".to_string(),
                     publisher_shortname: Some("OAE".to_string()),
@@ -469,6 +470,7 @@ mod tests {
   "imprint": {
     "imprintName": "OA Editions Imprint",
     "imprintUrl": null,
+    "crossmarkDoi": null,
     "publisher": {
       "publisherName": "OA Editions",
       "publisherShortname": "OAE",